+++ js/data.js
@@ -1,0 +1,24 @@
+const createMessage = () =>
+  Array.from({ length: getRandomPositiveInteger(1, 2) }, () =>
+    getRandomArrayElement(MESSAGES)
+  ).join(' ');
+
+const createComment = () => ({
+  id: generateCommentId(),
+  avatar: `img/avatar-${getRandomPositiveInteger(1, AVATAR)}.svg`,
+  message: createMessage(),
+  name: getRandomArrayElement(NAMES),
+});
+
+const createPicture = (index) => ({
+  id: index,
+  url: `photos/${index}.jpg`,
+  description: getRandomArrayElement(DESCRIPTIONS),
+  likes: getRandomPositiveInteger(LIKES_MIN, LIKES_MAX),
+  comments: Array.from(
+    { length: getRandomPositiveInteger(0, COMMENTS) },
+    createComment
+  ),
+});
+
+export {createMessage, createComment, createPicture};
