+++ js/main.js
@@ -1,3 +1,6 @@
+import {getRandomPositiveInteger} from './util.js';
+import {createMessage, createComment, createPicture} from './data.js';
+
 const POSTED_PICTURES = 25;
 const LIKES_MIN = 15;
 const LIKES_MAX = 200;
@@ -10,16 +13,6 @@ const MESSAGES = [
 const DESCRIPTIONS = ['Это мой лучший кадр!', 'Я будущий популярный блогер!'];
 const NAMES = ['Оксана', 'Матвей', 'Виктория', 'Тимофей', 'Юрий'];
 
-const getRandomPositiveInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const getRandomArrayElement = (array) =>
-  array[getRandomPositiveInteger(0, array.length - 1)];
-
 const createIdGenerator = () => {
   let lastGeneratedId = 0;
   return () => {
@@ -29,29 +22,6 @@ const createIdGenerator = () => {
 };
 
 const generateCommentId = createIdGenerator();
-
-const createMessage = () =>
-  Array.from({ length: getRandomPositiveInteger(1, 2) }, () =>
-    getRandomArrayElement(MESSAGES)
-  ).join(' ');
-
-const createComment = () => ({
-  id: generateCommentId(),
-  avatar: `img/avatar-${getRandomPositiveInteger(1, AVATAR)}.svg`,
-  message: createMessage(),
-  name: getRandomArrayElement(NAMES),
-});
-
-const createPicture = (index) => ({
-  id: index,
-  url: `photos/${index}.jpg`,
-  description: getRandomArrayElement(DESCRIPTIONS),
-  likes: getRandomPositiveInteger(LIKES_MIN, LIKES_MAX),
-  comments: Array.from(
-    { length: getRandomPositiveInteger(0, COMMENTS) },
-    createComment
-  ),
-});
 
 const getPictures = () =>
   Array.from({ length: POSTED_PICTURES }, (_, pictureIndex) =>
